refactor(cosmic-swingset): move install bundle logic From block-manager into launch-chain and chain-main
