+++ packages/cosmic-swingset/src/block-manager.js
@@ -2,10 +2,8 @@
 /* global process */
 import anylogger from 'anylogger';
 
-import { makeMarshal } from '@endo/marshal';
 import { assert, details as X } from '@agoric/assert';
 import { BridgeId as BRIDGE_ID } from '@agoric/internal';
-import { makeStoredSubscriber, makePublishKit } from '@agoric/notifier';
 import * as ActionType from './action-types.js';
 import { parseParams } from './params.js';
 
@@ -24,6 +22,7 @@ export default function makeBlockManager({
   actionQueue,
   deliverInbound,
   doBridgeInbound,
+  installBundle,
   bootstrapBlock,
   beginBlock,
   endBlock,
@@ -31,8 +30,6 @@ export default function makeBlockManager({
   saveChainState,
   saveOutsideState,
   savedHeight,
-  validateAndInstallBundle,
-  installationStorageNode = undefined,
   verboseBlocks = false,
 }) {
   let computedHeight = savedHeight;
@@ -41,9 +38,6 @@ export default function makeBlockManager({
   let latestParams;
   let beginBlockAction;
 
-  /** @type {PublishKit<unknown>['publisher'] | undefined} */
-  let installationPublisher;
-
   async function processAction(action) {
     const start = Date.now();
     const finish = res => {
@@ -51,16 +45,6 @@ export default function makeBlockManager({
       runTime += Date.now() - start;
       return res;
     };
-
-    if (
-      installationPublisher === undefined &&
-      installationStorageNode !== undefined
-    ) {
-      const marshaller = makeMarshal();
-      const { publisher, subscriber } = makePublishKit();
-      makeStoredSubscriber(subscriber, installationStorageNode, marshaller);
-      installationPublisher = publisher;
-    }
 
     // console.error('Performing action', action);
     let p;
@@ -99,27 +83,7 @@ export default function makeBlockManager({
       }
 
       case ActionType.INSTALL_BUNDLE: {
-        p = (async () => {
-          const bundle = JSON.parse(action.bundle);
-          harden(bundle);
-
-          const error = await validateAndInstallBundle(bundle).then(
-            () => null,
-            (/** @type {unknown} */ errorValue) => errorValue,
-          );
-
-          const { endoZipBase64Sha512 } = bundle;
-
-          if (installationPublisher !== undefined) {
-            installationPublisher.publish(
-              harden({
-                endoZipBase64Sha512,
-                installed: error === null,
-                error,
-              }),
-            );
-          }
-        })();
+        p = installBundle(action.bundle);
         break;
       }
